Add useWriterLock option to the Repo.
useWriterLock ensures events become visible with monotocially increasing
journalIndex, so that streams do not skip events. This is not guaranteed
with when there are several writers (many nodes) currently.

This is currently based on SELECT FOR UPDATE on a single row in a
different table. The reason for this is to ensure that only the event
writers execute serially, not all writes to the journal table (event
deletions and so on).